ath6kl: use linaro's generic wext driver
Unfortunately the modifications Samsung have made to net/wireless
and the ath6kl driver for their tablets have broken support for
the nl80211 interface. Until this can be sorted out, we need to
rely on a wext driver, and the current one is not functional (nor
used by any device).